Rollback spliten editor types
