--- luda-editor/luda-editor-client/src/app/types/image_filename_object.rs
@@ -1,19 +1,0 @@
-use super::CharacterPoseEmotion;
-
-#[derive(Debug, Clone)]
-pub struct ImageFilenameObject {
-    pub character: String,
-    pub pose: String,
-    pub emotion: String,
-    pub url: String,
-}
-
-impl ImageFilenameObject {
-    pub fn into_character_pose_emotion(&self) -> CharacterPoseEmotion {
-        CharacterPoseEmotion(
-            self.character.clone(),
-            self.pose.clone(),
-            self.emotion.clone(),
-        )
-    }
-}
